ringhash: move out of the xds package
fixes https://github.com/grpc/grpc-go/issues/8213
