+++ balancer/ringhash/util.go
